+++ CHANGELOG.md
@@ -11,6 +11,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - feat: Initial commit
 - feat: Add python SDK skeleton
+- feat: Add python SDK command auditor
 - feat: Add python SDK command builder
 - feat: Add python SDK command checker
 - feat: Add python SDK command documenter